refactor: 시간 관련 필드들 공통 클래스로 묶음 [CommonTimeEntity]
